+++ packages/react/src/components/tree-view/index.ts
@@ -1,11 +1,15 @@
 export type {
+  CheckedChangeDetails as TreeViewCheckedChangeDetails,
   ExpandedChangeDetails as TreeViewExpandedChangeDetails,
   FocusChangeDetails as TreeViewFocusChangeDetails,
-  SelectionChangeDetails as TreeViewSelectionChangeDetails,
-  LoadChildrenDetails as TreeViewLoadChildrenDetails,
   LoadChildrenCompleteDetails as TreeViewLoadChildrenCompleteDetails,
-  NodeState as TreeViewNodeState,
+  LoadChildrenDetails as TreeViewLoadChildrenDetails,
+  LoadChildrenErrorDetails as TreeViewLoadChildrenErrorDetails,
   NodeProps as TreeViewNodeProps,
+  NodeState as TreeViewNodeState,
+  RenameCompleteDetails as TreeViewRenameCompleteDetails,
+  RenameStartDetails as TreeViewRenameStartDetails,
+  SelectionChangeDetails as TreeViewSelectionChangeDetails,
 } from '@zag-js/tree-view'
 export { createFileTreeCollection, createTreeCollection, type TreeCollection, type TreeNode } from '../collection'
 export { TreeViewBranch, type TreeViewBranchBaseProps, type TreeViewBranchProps } from './tree-view-branch'
@@ -78,6 +82,11 @@ export {
   type TreeViewNodeCheckboxIndicatorBaseProps,
   type TreeViewNodeCheckboxIndicatorProps,
 } from './tree-view-node-checkbox-indicator'
+export {
+  TreeViewNodeRenameInput,
+  type TreeViewNodeRenameInputBaseProps,
+  type TreeViewNodeRenameInputProps,
+} from './tree-view-node-rename-input'
 export { treeViewAnatomy } from './tree-view.anatomy'
 export { useTreeView, type UseTreeViewProps, type UseTreeViewReturn } from './use-tree-view'
 export { useTreeViewContext, type UseTreeViewContext } from './use-tree-view-context'