+++ packages/solid/src/components/tree-view/index.tsx
@@ -1,11 +1,15 @@
 export type {
+  CheckedChangeDetails as TreeViewCheckedChangeDetails,
   ExpandedChangeDetails as TreeViewExpandedChangeDetails,
   FocusChangeDetails as TreeViewFocusChangeDetails,
-  SelectionChangeDetails as TreeViewSelectionChangeDetails,
-  LoadChildrenDetails as TreeViewLoadChildrenDetails,
   LoadChildrenCompleteDetails as TreeViewLoadChildrenCompleteDetails,
-  NodeState as TreeViewNodeState,
+  LoadChildrenDetails as TreeViewLoadChildrenDetails,
+  LoadChildrenErrorDetails as TreeViewLoadChildrenErrorDetails,
   NodeProps as TreeViewNodeProps,
+  NodeState as TreeViewNodeState,
+  RenameCompleteDetails as TreeViewRenameCompleteDetails,
+  RenameStartDetails as TreeViewRenameStartDetails,
+  SelectionChangeDetails as TreeViewSelectionChangeDetails,
 } from '@zag-js/tree-view'
 export { createFileTreeCollection, createTreeCollection, type TreeCollection, type TreeNode } from '../collection'
 export { TreeViewBranch, type TreeViewBranchBaseProps, type TreeViewBranchProps } from './tree-view-branch'
@@ -58,6 +62,11 @@ export {
   type TreeViewNodeCheckboxIndicatorBaseProps,
   type TreeViewNodeCheckboxIndicatorProps,
 } from './tree-view-node-checkbox-indicator'
+export {
+  TreeViewNodeRenameInput,
+  type TreeViewNodeRenameInputBaseProps,
+  type TreeViewNodeRenameInputProps,
+} from './tree-view-node-rename-input'
 export { TreeViewNodeContext, type TreeViewNodeContextProps } from './tree-view-node-context'
 export {
   TreeViewNodeProvider,